refactor: join unpopulateNotificationsDB scripts
Move all the unpopulateNotificationsScript's functions into
unpopulateNotificationsDB file. This is because the functions aren't
used in another file except unpopulateNotificationsDB.
